+++ geotrek/tourism/tests/test_parsers.py
@@ -1402,7 +1402,7 @@ class InformationDeskGeotrekParserTests(GeotrekParserTestMixin, TestCase):
 
 class TestInformationDeskOpenStreetMapParser(InformationDeskOpenStreetMapParser):
     type = "Foo"
-    tags = {"amenity": "ranger_station"}
+    tags = [{"amenity": "ranger_station"}]
 
     def filter_name(self, src, val):
         if val:
@@ -1442,46 +1442,46 @@ class OpenStreetMapParserTests(TestCase):
         self.assertEqual(self.objects.count(), 4)
 
     def test_get_tag_info_existing_tag_OSM(self):
-        information_desk = self.objects.get(eid=1)
+        information_desk = self.objects.get(eid="1")
         self.assertEqual(information_desk.phone, "0754347899")
 
-        information_desk2 = self.objects.get(eid=2)
+        information_desk2 = self.objects.get(eid="2")
         self.assertEqual(information_desk2.phone, "0754347899")
 
     def test_get_tag_info_no_tag_OSM(self):
-        information_desk = self.objects.get(eid=3)
+        information_desk = self.objects.get(eid="3")
         self.assertEqual(information_desk.phone, None)
 
     def test_InformationDesk_street_filter_housenumber_and_street_OSM(self):
-        information_desk = self.objects.get(eid=1)
+        information_desk = self.objects.get(eid="1")
         self.assertEqual(information_desk.street, "5 rue des chênes")
 
     def test_InformationDesk_street_filter_street_OSM(self):
-        information_desk = self.objects.get(eid=2)
+        information_desk = self.objects.get(eid="2")
         self.assertEqual(information_desk.street, "rue des chênes")
 
     def test_InformationDesk_street_filter_None_OSM(self):
-        information_desk = self.objects.get(eid=3)
+        information_desk = self.objects.get(eid="3")
         self.assertEqual(information_desk.street, None)
 
     def test_geom_point_to_point_OSM(self):
-        information_desk = self.objects.get(eid=1)
+        information_desk = self.objects.get(eid="1")
         self.assertAlmostEqual(information_desk.geom.coords[0], 673775.5074406686)
         self.assertAlmostEqual(information_desk.geom.coords[1], 6260613.093389216)
 
     def test_geom_way_to_point_OSM(self):
-        information_desk = self.objects.get(eid=3)
+        information_desk = self.objects.get(eid="3")
         self.assertAlmostEqual(information_desk.geom.coords[0], 639380.854410392)
         self.assertAlmostEqual(information_desk.geom.coords[1], 6256494.451055847)
 
     def test_geom_relation_to_point_OSM(self):
-        information_desk = self.objects.get(eid=4)
+        information_desk = self.objects.get(eid="4")
         self.assertAlmostEqual(information_desk.geom.coords[0], -5898321.244682654)
         self.assertAlmostEqual(information_desk.geom.coords[1], 12807160.659235487)
 
     def test_flexible_fields(self):
-        information_desk = self.objects.get(eid=1)
+        information_desk = self.objects.get(eid="1")
         self.assertEqual(information_desk.name, "test")
 
-        information_desk2 = self.objects.get(eid=4)
+        information_desk2 = self.objects.get(eid="4")
         self.assertEqual(information_desk2.name, "test_flexible")